requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
google-generativeai==0.3.2
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
pillow>=10.0.0
